First draft of Working Around Iterators as a Generic Type Parameter
This is a snippet for the Haxe Cookbook related to the use of abstracts for having code that can compile in any target, given Haxe's limitation on generic type parameters. Generic type parameters only accept classes or enum instances, not structures, which excludes the possibility of using Iterator out of the box. However, by wrapping Iterator on an abstract, it's possible to work around this limitation.